+++ NEWS.md
@@ -1,3 +1,12 @@
+# tidychangepoint 1.0.1
+
+* Added preliminary support for ARIMA models. 
+* Added support for models from the `changepointGA` package.
+* Added support for models from the `segmented` package.
+* `fitness.wbs()` returns the MBIC penalty value.
+* Added `summary()` methods for `tidycpt`, `seg_cpt`, and `mod_cpt` objects.
+* Improved output from `print()` methods.
+
 # tidychangepoint 1.0.0
 
 * `mlb_diffs` is now a `tsibble` with various statistics.